clean: 前年度作成したufac用ロジックの移動
